pyproject.toml
[tool.poetry]
name = "chroma-crash"
version = "0.1.0"
description = ""
authors = ["urig <urig@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "chroma_crash"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
chromadb = "0.4.8"
google-cloud-logging = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
